moved gamedata.py
moved gamedata.py from GUI to data_container
